Changed domain structure
As I forgot that Scala 2 doesn't have top level definitions
