+++ examples/src/main/scala/io/chrisdavenport/shellfish/contacts/domain/flag.scala
